requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.6.0
Pillow>=8.0.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
tqdm>=4.60.0